requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.3
pymongo==4.6.1
scikit-learn==1.3.2
streamlit==1.31.0
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.3